adk-personal-concierge/requirements.txt
google-adk
faiss-cpu
sentence-transformers
numpy